+++ Bookong.Application/DTOs/CreateBookDto.cs
@@ -98,50 +98,4 @@ namespace Bookong.Application.DTOs
             }
         }
     }
-
-    public class NewAuthorDto
-    {
-        // Temporary negative id used on the client to reference this new item before persistence.
-        public int TempId { get; set; }
-        public string FirstName { get; set; } = string.Empty;
-        public string? MiddleName { get; set; }
-        public string LastName { get; set; } = string.Empty;
-    }
-
-    public class NewNameDto
-    {
-        public int TempId { get; set; }
-        public string Name { get; set; } = string.Empty;
-    }
-
-    public class NewPublisherDto
-    {
-        public int TempId { get; set; }
-        public string Name { get; set; } = string.Empty;
-    }
-
-    public class NewWarehouseDto
-    {
-        public int TempId { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public string? Street { get; set; }
-        public string? Number { get; set; }
-        public string? City { get; set; }
-        public string? ZipCode { get; set; }
-
-    }
-
-    // New DTOs for form option lists
-    public class SelectItemDto { public int Id { get; set; } public string Name { get; set; } = string.Empty; }
-
-    public class FormOptionsDto
-    {
-        public List<SelectItemDto> Authors { get; set; } = new();
-        public List<SelectItemDto> Genres { get; set; } = new();
-        public List<SelectItemDto> Kinds { get; set; } = new();
-        public List<SelectItemDto> Periods { get; set; } = new();
-        public List<SelectItemDto> Publishers { get; set; } = new();
-        public List<SelectItemDto> Warehouses { get; set; } = new();
-
-    }
 }
