+++ Bookong.Application/DTOs/NewPublisherDto.cs
@@ -1,0 +1,8 @@
+namespace Bookong.Application.DTOs
+{
+ public class NewPublisherDto
+ {
+ public int TempId { get; set; }
+ public string Name { get; set; } = string.Empty;
+ }
+}
