+++ Bookong.Application/DTOs/SelectItemDto.cs
@@ -1,0 +1,4 @@
+namespace Bookong.Application.DTOs
+{
+ public class SelectItemDto { public int Id { get; set; } public string Name { get; set; } = string.Empty; }
+}
